# contracts/openapi.yaml
# Must match what FastAPI generates (app.openapi()); test_contract.py fails on drift.
openapi: 3.1.0
info:
  title: pro-wms API
  # Must equal app.__version__; api/tests/test_contract.py fails on drift.
  version: 0.2.0
  description: >-
    Contract shared by the CLI and the web console; the service is the FastAPI app in api/.
    This is a curated subset (paths, methods, identity), not a dump of the generated
    schema. api/tests/test_contract.py keeps it from drifting from the running app.
servers:
  - url: http://127.0.0.1:8080
paths:
  /health/ready:
    get:
      summary: Storage and schema readiness
      responses:
        "200": { description: ready }
        "503": { description: unavailable }
  /outbounds/{outbound_id}/cancel:
    post:
      summary: Cancel draft or allocated outbound and release reservations
      parameters:
        - { name: outbound_id, in: path, required: true, schema: { type: string } }
        - $ref: "#/components/parameters/UserHeader"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200": { description: cancelled }
        "409": { description: already waved, shipped or cancelled }
  /health:
    get:
      summary: Liveness probe
      responses:
        "200": { description: ok }
  /snapshot:
    get:
      summary: Full kernel snapshot (warehouses, stock, documents) for the admin console
      responses:
        "200": { description: to_dict payload }
  /ledger:
    get:
      summary: Ledger rows plus on-hand quantity for one SKU
      parameters:
        - { name: sku, in: query, required: true, schema: { type: string } }
        - { name: warehouse, in: query, required: false, schema: { type: string } }
      responses:
        "200": { description: on_hand (physical), reserved, available plus ledger rows }
  /seed/demo:
    post:
      summary: Explicit demo mode only; PostgreSQL initialization requires an empty database
      parameters:
        - $ref: "#/components/parameters/UserHeader"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200": { description: seeded }
  /inbounds/{inbound_id}/receive:
    post:
      summary: Book goods onto the dock (draft -> receiving)
      parameters:
        - { name: inbound_id, in: path, required: true, schema: { type: string } }
        - $ref: "#/components/parameters/UserHeader"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200": { description: receiving }
        "404": { description: unknown inbound or lot }
        "409": { description: illegal transition }
  /inbounds/{inbound_id}/putaway:
    post:
      summary: Move received goods to a storage bin (receiving -> closed)
      parameters:
        - { name: inbound_id, in: path, required: true, schema: { type: string } }
        - $ref: "#/components/parameters/UserHeader"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to]
              properties:
                to: { type: string, description: target location id }
      responses:
        "200": { description: closed }
        "409": { description: illegal transition or short stock }
  /outbounds/{outbound_id}/allocate:
    post:
      summary: Reserve stock with FIFO / FEFO (draft -> allocated)
      parameters:
        - { name: outbound_id, in: path, required: true, schema: { type: string } }
        - $ref: "#/components/parameters/UserHeader"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                strategy: { type: string, enum: [fifo, fefo] }
      responses:
        "200": { description: "allocated; meta.allocations[] carries version_seen" }
        "409": { description: version conflict or insufficient stock }
  /waves:
    post:
      summary: Group allocated outbounds into a pick wave (supervisor only)
      parameters:
        - $ref: "#/components/parameters/UserHeader"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [outbounds]
              properties:
                outbounds:
                  type: array
                  items: { type: string }
      responses:
        "200": { description: created }
        "400": { description: empty, duplicate or cross-warehouse outbound list }
        "403": { description: supervisor identity required }
  /waves/{wave_id}/pick:
    post:
      summary: Complete a wave (waved -> shipped)
      parameters:
        - { name: wave_id, in: path, required: true, schema: { type: string } }
        - $ref: "#/components/parameters/UserHeader"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200": { description: shipped }
        "409": { description: wave is not pickable }
  /stocktakes/{stocktake_id}/approve:
    post:
      summary: Approve a submitted stocktake (supervisor only)
      parameters:
        - { name: stocktake_id, in: path, required: true, schema: { type: string } }
        - $ref: "#/components/parameters/UserHeader"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200": { description: approved }
        "403": { description: supervisor identity required }
  /race:
    post:
      summary: Reproduce concurrent allocation conflicts (demo / debug endpoint)
      parameters:
        - $ref: "#/components/parameters/UserHeader"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                sku: { type: string }
                warehouse: { type: string }
                workers: { type: integer, minimum: 1, maximum: 32 }
      responses:
        "200": { description: one winner plus conflicts }
        "422": { description: workers outside 1..32 }
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: >-
        Reuse for retries of the same operation, resolved actor and payload. A different
        request with the same key returns 409. Only successful transactions are recorded.
        Durable in PostgreSQL; process-local in memory mode. Not authentication.
      schema: { type: string, minLength: 1, maxLength: 128 }
    UserHeader:
      name: X-User
      in: header
      required: false
      description: >-
        Acting identity (operator / supervisor / admin). Defaults to the least privileged
        `operator`, so privileged endpoints return 403 unless an identity is declared.
        `?as=` and a body `user` field are accepted as fallbacks.
      schema: { type: string }